pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diabetic-foot-agent"
version = "0.1.0"
description = "A course-ready prototype for diabetic foot risk screening and TCM-assisted management."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.0",
  "pillow>=10.0",
  "streamlit>=1.34",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
